Move BigInt code to BigIntUtils (#413)
